Organize the code and fix the web version.
