+++ Cargo.toml
@@ -30,13 +30,14 @@ Video. Images. Sound. Post. VISP is an open source multimedia editor built entir
 
 [dependencies]
 dioxus-router = { version = "0.5.0-alpha.0" }
-gstreamer = "0.22.2"
-gstreamer-app = "0.22.0"
+
 web-sys = { version = "0.3.68", features = ["Blob", "FilePropertyBag", "Url", "File"] }
 
 [target.'cfg(target_family = "windows")'.dependencies]
 dioxus = { version = "0.5.0-alpha.0", features = ["desktop"] }
 platform-dirs = "0.3.0"
+gstreamer = "0.22.2"
+gstreamer-app = "0.22.0"
 
 [target.'cfg(target_family = "wasm")'.dependencies]
 dioxus = { version = "0.5.0-alpha.0", features = ["web"] }
